EcoShop-main/requirements.txt
Flask
flask_bcrypt
numpy
Pillow
pymongo
tensorflow
transformers
langchain